add support for continuations outside luau
