+++ src/state/extra.rs
@@ -13,6 +13,7 @@ use crate::error::Result;
 use crate::state::RawLua;
 use crate::stdlib::StdLib;
 use crate::types::{AppData, ReentrantMutex, XRc};
+
 use crate::userdata::RawUserDataRegistry;
 use crate::util::{get_internal_metatable, push_internal_userdata, TypeKey, WrappedFailure};
 
@@ -98,6 +99,9 @@ pub(crate) struct ExtraData {
 
     // Values currently being yielded from Lua.yield()
     pub(super) yielded_values: Option<MultiValue>,
+
+    #[cfg(all(not(feature = "luau"), not(feature = "lua51"), not(feature = "luajit")))]
+    pub(super) yield_continuation: bool,
 }
 
 impl Drop for ExtraData {
@@ -201,6 +205,8 @@ impl ExtraData {
             #[cfg(feature = "luau")]
             running_gc: false,
             yielded_values: None,
+            #[cfg(all(not(feature = "luau"), not(feature = "lua51"), not(feature = "luajit")))]
+            yield_continuation: false,
         }));
 
         // Store it in the registry
