+++ src/state/raw.rs
@@ -11,7 +11,9 @@ use crate::chunk::ChunkMode;
 use crate::error::{Error, Result};
 use crate::function::Function;
 use crate::memory::{MemoryState, ALLOCATOR};
-use crate::state::util::{callback_error_ext, callback_error_ext_yieldable, ref_stack_pop};
+#[cfg(not(feature = "luau"))]
+use crate::state::util::callback_error_ext;
+use crate::state::util::{callback_error_ext_yieldable, ref_stack_pop};
 use crate::stdlib::StdLib;
 use crate::string::String;
 use crate::table::Table;
@@ -22,8 +24,10 @@ use crate::types::{
     MaybeSend, ReentrantMutex, RegistryKey, ValueRef, XRc,
 };
 
-#[cfg(feature = "luau")]
-use crate::types::{LuauContinuation, LuauContinuationUpvalue};
+#[cfg(all(not(feature = "lua51"), not(feature = "luajit")))]
+use crate::types::Continuation;
+#[cfg(all(not(feature = "lua51"), not(feature = "luajit")))]
+use crate::types::ContinuationUpvalue;
 
 use crate::userdata::{
     init_userdata_metatable, AnyUserData, MetaMethod, RawUserDataRegistry, UserData, UserDataRegistry,
@@ -201,8 +205,8 @@ impl RawLua {
                 init_internal_metatable::<XRc<UnsafeCell<ExtraData>>>(state, None)?;
                 init_internal_metatable::<Callback>(state, None)?;
                 init_internal_metatable::<CallbackUpvalue>(state, None)?;
-                #[cfg(feature = "luau")]
-                init_internal_metatable::<LuauContinuationUpvalue>(state, None)?;
+                #[cfg(all(not(feature = "lua51"), not(feature = "luajit")))]
+                init_internal_metatable::<ContinuationUpvalue>(state, None)?;
                 #[cfg(not(feature = "luau"))]
                 init_internal_metatable::<HookCallback>(state, None)?;
                 #[cfg(feature = "async")]
@@ -1222,56 +1226,105 @@ impl RawLua {
     }
 
     // Creates a Function out of a Callback and a continuation containing a 'static Fn.
-    #[cfg(feature = "luau")]
-    pub(crate) fn create_callback_with_luau_continuation(
+    //
+    // In Luau, uses pushcclosurek
+    //
+    // In Lua 5.2/5.3/5.4/JIT, makes a normal function that then yields to the continuation via yieldk
+    #[cfg(all(not(feature = "lua51"), not(feature = "luajit")))]
+    pub(crate) fn create_callback_with_continuation(
         &self,
         func: Callback,
-        cont: LuauContinuation,
+        cont: Continuation,
     ) -> Result<Function> {
-        unsafe extern "C-unwind" fn call_callback(state: *mut ffi::lua_State) -> c_int {
-            let upvalue = get_userdata::<LuauContinuationUpvalue>(state, ffi::lua_upvalueindex(1));
-            callback_error_ext_yieldable(state, (*upvalue).extra.get(), true, |extra, nargs| {
-                // Lua ensures that `LUA_MINSTACK` stack spaces are available (after pushing arguments)
-                // The lock must be already held as the callback is executed
-                let rawlua = (*extra).raw_lua();
-                match (*upvalue).data {
-                    Some(ref func) => (func.0)(rawlua, nargs),
-                    None => Err(Error::CallbackDestructed),
-                }
-            })
-        }
-
-        unsafe extern "C-unwind" fn cont_callback(state: *mut ffi::lua_State, status: c_int) -> c_int {
-            let upvalue = get_userdata::<LuauContinuationUpvalue>(state, ffi::lua_upvalueindex(1));
-            callback_error_ext_yieldable(state, (*upvalue).extra.get(), true, |extra, nargs| {
-                // Lua ensures that `LUA_MINSTACK` stack spaces are available (after pushing arguments)
-                // The lock must be already held as the callback is executed
-                let rawlua = (*extra).raw_lua();
-                match (*upvalue).data {
-                    Some(ref func) => (func.1)(rawlua, nargs, status),
-                    None => Err(Error::CallbackDestructed),
-                }
-            })
-        }
-
-        let state = self.state();
-        unsafe {
-            let _sg = StackGuard::new(state);
-            check_stack(state, 4)?;
-
-            let func = Some((func, cont));
-            let extra = XRc::clone(&self.extra);
-            let protect = !self.unlikely_memory_error();
-            push_internal_userdata(state, LuauContinuationUpvalue { data: func, extra }, protect)?;
-            if protect {
-                protect_lua!(state, 1, 1, fn(state) {
-                    ffi::lua_pushcclosurec(state, call_callback, cont_callback, 1);
-                })?;
-            } else {
-                ffi::lua_pushcclosurec(state, call_callback, cont_callback, 1);
+        #[cfg(feature = "luau")]
+        {
+            unsafe extern "C-unwind" fn call_callback(state: *mut ffi::lua_State) -> c_int {
+                let upvalue = get_userdata::<ContinuationUpvalue>(state, ffi::lua_upvalueindex(1));
+                callback_error_ext_yieldable(state, (*upvalue).extra.get(), true, |extra, nargs| {
+                    // Lua ensures that `LUA_MINSTACK` stack spaces are available (after pushing arguments)
+                    // The lock must be already held as the callback is executed
+                    let rawlua = (*extra).raw_lua();
+                    match (*upvalue).data {
+                        Some(ref func) => (func.0)(rawlua, nargs),
+                        None => Err(Error::CallbackDestructed),
+                    }
+                })
             }
 
-            Ok(Function(self.pop_ref()))
+            unsafe extern "C-unwind" fn cont_callback(state: *mut ffi::lua_State, status: c_int) -> c_int {
+                let upvalue = get_userdata::<ContinuationUpvalue>(state, ffi::lua_upvalueindex(1));
+                callback_error_ext_yieldable(state, (*upvalue).extra.get(), true, |extra, nargs| {
+                    // Lua ensures that `LUA_MINSTACK` stack spaces are available (after pushing arguments)
+                    // The lock must be already held as the callback is executed
+                    let rawlua = (*extra).raw_lua();
+                    match (*upvalue).data {
+                        Some(ref func) => (func.1)(rawlua, nargs, status),
+                        None => Err(Error::CallbackDestructed),
+                    }
+                })
+            }
+
+            let state = self.state();
+            unsafe {
+                let _sg = StackGuard::new(state);
+                check_stack(state, 4)?;
+
+                let func = Some((func, cont));
+                let extra = XRc::clone(&self.extra);
+                let protect = !self.unlikely_memory_error();
+                push_internal_userdata(state, ContinuationUpvalue { data: func, extra }, protect)?;
+                if protect {
+                    protect_lua!(state, 1, 1, fn(state) {
+                        ffi::lua_pushcclosurec(state, call_callback, cont_callback, 1);
+                    })?;
+                } else {
+                    ffi::lua_pushcclosurec(state, call_callback, cont_callback, 1);
+                }
+
+                Ok(Function(self.pop_ref()))
+            }
+        }
+
+        #[cfg(not(feature = "luau"))]
+        {
+            unsafe extern "C-unwind" fn call_callback(state: *mut ffi::lua_State) -> c_int {
+                let upvalue = get_userdata::<ContinuationUpvalue>(state, ffi::lua_upvalueindex(1));
+                callback_error_ext_yieldable(state, (*upvalue).extra.get(), true, |extra, nargs| {
+                    // Lua ensures that `LUA_MINSTACK` stack spaces are available (after pushing arguments)
+                    // The lock must be already held as the callback is executed
+                    let rawlua = (*extra).raw_lua();
+                    match (*upvalue).data {
+                        Some((ref func, _)) => match func(rawlua, nargs) {
+                            Ok(r) => {
+                                (*extra).yield_continuation = true;
+                                Ok(r)
+                            }
+                            Err(e) => Err(e),
+                        },
+                        None => Err(Error::CallbackDestructed),
+                    }
+                })
+            }
+
+            let state = self.state();
+            unsafe {
+                let _sg = StackGuard::new(state);
+                check_stack(state, 4)?;
+
+                let func = Some((func, cont));
+                let extra = XRc::clone(&self.extra);
+                let protect = !self.unlikely_memory_error();
+                push_internal_userdata(state, ContinuationUpvalue { data: func, extra }, protect)?;
+                if protect {
+                    protect_lua!(state, 1, 1, fn(state) {
+                        ffi::lua_pushcclosure(state, call_callback, 1);
+                    })?;
+                } else {
+                    ffi::lua_pushcclosure(state, call_callback, 1);
+                }
+
+                Ok(Function(self.pop_ref()))
+            }
         }
     }
