+++ src/state/util.rs
@@ -9,6 +9,9 @@ use crate::error::{Error, Result};
 use crate::state::{ExtraData, RawLua};
 use crate::util::{self, check_stack, get_internal_metatable, WrappedFailure};
 
+#[cfg(all(not(feature = "lua51"), not(feature = "luajit"), not(feature = "luau")))]
+use crate::{types::ContinuationUpvalue, util::get_userdata};
+
 struct StateGuard<'a>(&'a RawLua, *mut ffi::lua_State);
 
 impl<'a> StateGuard<'a> {
@@ -112,6 +115,8 @@ where
         Ok(Ok(r)) => {
             // Ensure yielded values are cleared
             take(&mut extra.as_mut().unwrap_unchecked().yielded_values);
+            #[cfg(all(not(feature = "luau"), not(feature = "lua51"), not(feature = "luajit")))]
+            take(&mut extra.as_mut().unwrap_unchecked().yield_continuation);
 
             // Return unused `WrappedFailure` to the pool
             prealloc_failure.release(state, extra);
@@ -188,6 +193,9 @@ where
             let raw = extra.as_ref().unwrap_unchecked().raw_lua();
             let values = take(&mut extra.as_mut().unwrap_unchecked().yielded_values);
 
+            #[cfg(all(not(feature = "luau"), not(feature = "lua51"), not(feature = "luajit")))]
+            let yield_cont = take(&mut extra.as_mut().unwrap_unchecked().yield_continuation);
+
             if let Some(values) = values {
                 if raw.state() == state {
                     // Edge case: main thread is being yielded
@@ -217,6 +225,67 @@ where
                             }
                             ffi::lua_xmove(raw.state(), state, nargs);
                         }
+
+                        #[cfg(all(not(feature = "luau"), not(feature = "lua51"), not(feature = "luajit")))]
+                        {
+                            // Yield to a continuation. Unlike luau, we need to do this manually and on the
+                            // fly using a yieldk call
+                            if yield_cont {
+                                // On Lua 5.2, status and ctx are not present, so use 0 as status for
+                                // compatibility
+                                #[cfg(feature = "lua52")]
+                                unsafe extern "C-unwind" fn cont_callback(
+                                    state: *mut ffi::lua_State,
+                                ) -> c_int {
+                                    let upvalue =
+                                        get_userdata::<ContinuationUpvalue>(state, ffi::lua_upvalueindex(1));
+                                    callback_error_ext_yieldable(
+                                        state,
+                                        (*upvalue).extra.get(),
+                                        true,
+                                        |extra, nargs| {
+                                            // Lua ensures that `LUA_MINSTACK` stack spaces are available
+                                            // (after pushing arguments)
+                                            // The lock must be already held as the callback is executed
+                                            let rawlua = (*extra).raw_lua();
+                                            match (*upvalue).data {
+                                                Some(ref func) => (func.1)(rawlua, nargs, 0),
+                                                None => Err(Error::CallbackDestructed),
+                                            }
+                                        },
+                                    )
+                                }
+
+                                // Lua 5.3/5.4 case
+                                #[cfg(not(feature = "lua52"))]
+                                unsafe extern "C-unwind" fn cont_callback(
+                                    state: *mut ffi::lua_State,
+                                    status: c_int,
+                                    _ctx: ffi::lua_KContext,
+                                ) -> c_int {
+                                    let upvalue =
+                                        get_userdata::<ContinuationUpvalue>(state, ffi::lua_upvalueindex(1));
+                                    callback_error_ext_yieldable(
+                                        state,
+                                        (*upvalue).extra.get(),
+                                        true,
+                                        |extra, nargs| {
+                                            // Lua ensures that `LUA_MINSTACK` stack spaces are available
+                                            // (after pushing arguments)
+                                            // The lock must be already held as the callback is executed
+                                            let rawlua = (*extra).raw_lua();
+                                            match (*upvalue).data {
+                                                Some(ref func) => (func.1)(rawlua, nargs, status),
+                                                None => Err(Error::CallbackDestructed),
+                                            }
+                                        },
+                                    )
+                                }
+
+                                return ffi::lua_yieldc(state, nargs, cont_callback);
+                            }
+                        }
+
                         return ffi::lua_yield(state, nargs);
                     }
                     Err(err) => {
