+++ src/util/types.rs
@@ -3,8 +3,8 @@ use std::os::raw::c_void;
 
 use crate::types::{Callback, CallbackUpvalue};
 
-#[cfg(feature = "luau")]
-use crate::types::LuauContinuationUpvalue;
+#[cfg(all(not(feature = "lua51"), not(feature = "luajit")))]
+use crate::types::ContinuationUpvalue;
 
 #[cfg(feature = "async")]
 use crate::types::{AsyncCallback, AsyncCallbackUpvalue, AsyncPollUpvalue};
@@ -37,12 +37,12 @@ impl TypeKey for CallbackUpvalue {
     }
 }
 
-#[cfg(feature = "luau")]
-impl TypeKey for LuauContinuationUpvalue {
+#[cfg(all(not(feature = "lua51"), not(feature = "luajit")))]
+impl TypeKey for ContinuationUpvalue {
     #[inline(always)]
     fn type_key() -> *const c_void {
-        static LUAU_CONTINUATION_UPVALUE_TYPE_KEY: u8 = 0;
-        &LUAU_CONTINUATION_UPVALUE_TYPE_KEY as *const u8 as *const c_void
+        static CONTINUATION_UPVALUE_TYPE_KEY: u8 = 0;
+        &CONTINUATION_UPVALUE_TYPE_KEY as *const u8 as *const c_void
     }
 }
